BigDecimal for amounts

+++ src/main/java/com/pasarella/prestamos/business/model/request/BCustomer.java
@@ -5,6 +5,8 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigInteger;
+
 
 @Builder
 @Setter
@@ -14,7 +16,7 @@ public class BCustomer {
 
     private Long idCustomer;
     private String name;
-    private Integer document;
+    private BigInteger document;
     private String telephone;
     private String jobLocalName;
     private String gender;
